remove another game default watcher with one usage
